tested result of distribute_editions in editions tests and move hasTokens function to common balance inspector file
