+++ contracts/test/fa2-balance-inspector.ts
@@ -13,3 +13,16 @@ export type QueryBalances = (
     fa2: Contract,
     requests: BalanceOfRequest[]
 ) => Promise<BalanceOfResponse[]>;
+
+export async function hasTokens(requests: BalanceOfRequest[],
+    queryBalances : QueryBalances,
+    nft : Contract 
+): Promise<boolean[]> {
+    const responses = await queryBalances(nft, requests);
+    const results = responses.map(r => {
+    if (r.balance.eq(1)) return true;
+    else if (r.balance.eq(0)) return false;
+    else throw new Error(`Invalid NFT balance ${r.balance}`);
+    });
+    return results;
+}
